+++ Changes.md
@@ -7,6 +7,8 @@ Version 3.4.0.5
         a node that contains the section name because CT has broken down the section content into two separate sub-sections
         and other core course formats only have the one section containing the section markup.  Therefore the solution effectively
         allows 'find_element_text' to find the hidden section name first within the 'labelroot' node that is determined.
+  2. When accessing a module through a single section page (not the navigation block) then ensure that you are returned
+     to the same single section page when using the breadcrumb.
 
 Version 3.4.0.4
   1. Fix CONTRIB-7393.